move create_output to own file
