+++ src/main.cpp
@@ -35,32 +35,6 @@
 #include "pattern_prevalence.hpp"
 #include "tile.hpp"
 
-Output create_output(const Model& model)
-{
-	Output output;
-	output._wave = Array3D<Bool>(model._width, model._height, model._num_patterns, true);
-	output._changes = Array2D<Bool>(model._width, model._height, false);
-
-	if (model._foundation != kInvalidIndex) {
-		for (const auto x : irange(model._width)) {
-			for (const auto t : irange(model._num_patterns)) {
-				if (t != model._foundation) {
-					output._wave.set(x, model._height - 1, t, false);
-				}
-			}
-			output._changes.set(x, model._height - 1, true);
-
-			for (const auto y : irange(model._height - 1)) {
-				output._wave.set(x, y, model._foundation, false);
-				output._changes.set(x, y, true);
-			}
-
-			while (model.propagate(&output));
-		}
-	}
-
-	return output;
-}
 
 void run_and_write(const Options& options, const std::string& name, const configuru::Config& config, const Model& model)
 {
